pyproject.toml
[project]
name = "sourcebox-article-drafter"
version = "0.1.0"
description = "Automated article idea drafter for SourceBox"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "python-dotenv>=1.2.1",
    "replicate>=1.0.7",
    "requests>=2.32.5",
]
